+++ forest/cli/src/cli/db_cmd.rs
@@ -1,7 +1,8 @@
 // Copyright 2019-2022 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use forest_cli_shared::{chain_path, cli::Config, db_path};
+use forest_cli_shared::{chain_path, cli::Config};
+use forest_db::db_engine::db_path;
 use log::error;
 use structopt::StructOpt;
