+++ forest/daemon/src/daemon.rs
@@ -8,11 +8,14 @@ use forest_auth::{create_token, generate_priv_key, ADMIN, JWT_IDENTIFIER};
 use forest_chain::ChainStore;
 use forest_chain_sync::consensus::SyncGossipSubmitter;
 use forest_chain_sync::ChainMuxer;
+use forest_cli_shared::chain_path;
 use forest_cli_shared::cli::{
     default_snapshot_dir, is_aria2_installed, snapshot_fetch, Client, Config, FOREST_VERSION_STRING,
 };
-use forest_cli_shared::{chain_path, db_path, open_db, Db};
-use forest_db::Store;
+use forest_db::{
+    db_engine::{db_path, open_db, Db},
+    Store,
+};
 use forest_genesis::{get_network_name_from_genesis, import_chain, read_genesis_header};
 use forest_key_management::ENCRYPTED_KEYSTORE_NAME;
 use forest_key_management::{KeyStore, KeyStoreConfig};
@@ -149,7 +152,7 @@ pub(super) async fn start(config: Config, detached: bool) -> anyhow::Result<Db> 
             "Prometheus server started at {}",
             config.client.metrics_address
         );
-        let db_directory = db_path(&chain_path(&config));
+        let db_directory = forest_db::db_engine::db_path(&chain_path(&config));
         let db = db.clone();
         services.spawn(async {
             forest_metrics::init_prometheus(prometheus_listener, db_directory, db)
